further SubscriptionController migrations
also: legacy removal
