--- grails-app/views/subscription/expected.gsp
@@ -1,1 +1,0 @@
-<g:render template="prev_exp_temp"  model="${[screen:'expected']}"/>
