--- grails-app/views/subscription/previous.gsp
@@ -1,1 +1,0 @@
-<g:render template="prev_exp_temp"  model="${[screen:'previous']}"/>
